moved the form modal outside as a seperate component
